requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
anthropic>=0.40.0
httpx>=0.25.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0.1
aiosqlite>=0.19.0
python-multipart>=0.0.6
pytz>=2023.3
websockets>=11.0



